pyproject.toml
[project]
name = "imageone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
]
